+++ index.html
@@ -233,6 +233,6 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-<script src="./js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
